Change game logic in backend (has not been tested)

+++ game/block.go
@@ -1,10 +1,10 @@
 package game
 
 type Block struct {
-	X float32 `json:"x"`
-	Y float32 `json:"y"`
-	w float32
-	h float32
+	X float64 `json:"x"`
+	Y float64 `json:"y"`
+	w float64
+	h float64
 }
 
 var blocks []*Block
